Allow specs to be viewed inline
As well as viewing files inline, this splits out directories from files,
and only shows those files that contain JSON objects.

This does mean that all the files will be read in, which could be time
and memory consuming, but the assumption is that this is being run in
directories which are primarily Vega/Vega-Lite specifications.