Pass errors from Vega-Embed to users and change default filename
Errors from Vega-Embed are now displayed to the user. It is not clear
if this provides much actionable information, but does indicate that
something went wrong.

File names for "export-to-svg/png" now use the specification file
name rather than the default prefix of 'vizualization'. There is no
attempt to remove suffixes (or directory elements, which may need
looking at in the future).