Add a horizontal scroll bar to the /display/...path-to-spec view
Rather than have the visualization disappear off the edge of the page,
with no way to adjust the view to see the rest, a horizontal scroll bar
will now appear (and the container element will be limited to the page
width). This does not happen for the vertical scroll bar (since the
one added by the browser is sufficient, even though it has a slightly
different user experience).

This only happens for individual visualizations displayed directly,
that is the /display/..path-to-spec end point. The inline views and
the drag-and-drop display do not provide scroll bars.