Change labelling of directories and display of inline plots
When displaying a visualization the path to the file is no-longer included
in the title bar, just the file name. This is because it duplicates
information already on the page, reduces the width of the title bar,
and better matches the drag-and-drop page (since there only the file
name, and not the path, is shown).

There have been style changes to how the inline plots are displayed,
to allow them to be horizontally scrolled.

A common stanza is used to indicate the current directory or file
and the path to the "parent".