Change the way the page listing specification works and is displayed
The inline version of the specification now appears to the right of the
list of specifications. The specification list has been changed from a
table to a more CSS-heavy "pseudo button" approach, hopefully reducing
some repetition and potential confusion.

An arrow SVG is emebedded for this page, to indicate an inline link,
and the SVG is from version 5.10.2 of the free version of fontawesome,
the arrow-alt-circle-right icon.